feat(index): Use flatbuffers instead of protobuf
Use flatbuffers instead of protobuf for storing table index.

+++ levels.go
@@ -1133,7 +1133,7 @@ type TableInfo struct {
 	Left        []byte
 	Right       []byte
 	KeyCount    uint64 // Number of keys in the table
-	EstimatedSz uint64
+	EstimatedSz uint32
 	IndexSz     int
 }
